refactor(images): split image view model responsibilities
